Delete some unused files. Move useComponentVisible into MoreOptions component
