+++ apps/modernization-ui/src/apps/page-builder/components/MoreOptions/useComponentVisible.tsx
